requirements.txt
Django==1.10.1
Markdown==2.6.6
MySQL-python==1.2.5
django-filter==0.14.0
djangorestframework==3.4.6
wsgiref==0.1.2
